Various CV tests and real offset values

+++ ut-robomaster/src/communication/beaglebone_communicator.hpp
@@ -72,7 +72,7 @@ public:
     /**
      * Identifier for the last turret data packet received, so we can check for new data
      */
-    uint8_t turretDataIndex = 0;
+    uint32_t turretDataIndex = 0;
 
 private:
     src::Drivers* drivers;
